Refactor: move prompt to embed txt
