--- agents/conversational_prompt.go
@@ -1,61 +1,0 @@
-//nolint:all
-package agents
-
-import (
-	"strings"
-
-	"github.com/tmc/langchaingo/prompts"
-	"github.com/tmc/langchaingo/tools"
-)
-
-const _defaultConversationalPrefix = `Assistant is a large language model trained by Meta.
-
-Assistant is designed to be able to assist with a wide range of tasks, from answering simple questions to providing in-depth explanations and discussions on a wide range of topics. As a language model, Assistant is able to generate human-like text based on the input it receives, allowing it to engage in natural-sounding conversations and provide responses that are coherent and relevant to the topic at hand.
-
-Assistant is constantly learning and improving, and its capabilities are constantly evolving. It is able to process and understand large amounts of text, and can use this knowledge to provide accurate and informative responses to a wide range of questions. Additionally, Assistant is able to generate its own text based on the input it receives, allowing it to engage in discussions and provide explanations and descriptions on a wide range of topics.
-
-Overall, Assistant is a powerful tool that can help with a wide range of tasks and provide valuable insights and information on a wide range of topics. Whether you need help with a specific question or just want to have a conversation about a particular topic, Assistant is here to assist.
-
-TOOLS:
-------
-
-Assistant has access to the following tools:
-
-{{.tool_descriptions}}`
-
-const _defaultConversationalFormatInstructions = `To use a tool, please use the following format:
-
-Thought: Do I need to use a tool? Yes
-Action: the action to take, should be one of [{{.tool_names}}]
-Action Input: the input to the action
-Observation: the result of the action
-
-When you have a response to say to the Human, or if you do not need to use a tool, you MUST use the format:
-
-Thought: Do I need to use a tool? No
-AI: [your response here]
-`
-
-const _defaultConversationalSuffix = `Begin!
-
-Previous conversation history:
-{{.history}}
-
-New input: {{.input}}
-
-Thought:{{.agent_scratchpad}}`
-
-func createConversationalPrompt(tools []tools.Tool, prefix, instructions, suffix string) prompts.PromptTemplate {
-	template := strings.Join([]string{prefix, instructions, suffix}, "\n\n")
-
-	return prompts.PromptTemplate{
-		Template:       template,
-		TemplateFormat: prompts.TemplateFormatGoTemplate,
-		InputVariables: []string{"input", "agent_scratchpad"},
-		PartialVariables: map[string]any{
-			"tool_names":        toolNames(tools),
-			"tool_descriptions": toolDescriptions(tools),
-			"history":           "",
-		},
-	}
-}
